feat(hosts): omp sessions
Oh My Pi writes pi's format under ~/.omp/agent/sessions and reuses pi's override
variables; the omp module is pi's reader over that root. Timestamp formatting moves to a
shared time module.

refs #24
