+++ crates/plannotator-tui-hosts/src/pi.rs
@@ -116,34 +116,12 @@ impl Entry {
 fn iso_timestamp(value: &Value) -> Option<String> {
     match value {
         Value::String(s) if !s.trim().is_empty() => Some(s.clone()),
-        Value::Number(n) => n.as_f64().filter(|f| f.is_finite() && *f >= 0.0).map(|ms| ms_to_iso(ms as u64)),
+        Value::Number(n) => n
+            .as_f64()
+            .filter(|f| f.is_finite() && *f >= 0.0)
+            .map(|ms| crate::time::iso_from_unix_ms(ms as u64)),
         _ => None,
     }
-}
-
-fn ms_to_iso(ms: u64) -> String {
-    let secs = ms / 1000;
-    let (year, month, day) = civil_from_days(secs / 86_400);
-    format!(
-        "{year:04}-{month:02}-{day:02}T{:02}:{:02}:{:02}.{:03}Z",
-        (secs / 3600) % 24,
-        (secs / 60) % 60,
-        secs % 60,
-        ms % 1000
-    )
-}
-
-/// Howard Hinnant's days-to-civil.
-fn civil_from_days(days: u64) -> (u64, u64, u64) {
-    let z = days + 719_468;
-    let era = z / 146_097;
-    let doe = z - era * 146_097;
-    let yoe = (doe - doe / 1460 + doe / 36524 - doe / 146_096) / 365;
-    let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
-    let mp = (5 * doy + 2) / 153;
-    let d = doy - (153 * mp + 2) / 5 + 1;
-    let m = if mp < 10 { mp + 3 } else { mp - 9 };
-    (yoe + era * 400 + u64::from(m <= 2), m, d)
 }
 
 /// The newest `n` messages on the active branch, newest first. The branch is the
